Move shortcut files to their own directory
