+++ app/classifier/tasks/shortcut/index.jsx
