+++ app/classifier/tasks/shortcut/shortcut-editor.spec.js
